Refactor and move add contract to it's own component - Tucker
