+++ components/UseYearComponent.js
@@ -1,76 +1,10 @@
 import Constants from 'expo-constants';
 import React, { useEffect } from 'react';
-import { StyleSheet, View } from "react-native";
-import { Card, Text, useTheme } from "react-native-paper";
+import { StyleSheet, TouchableOpacity, View } from "react-native";
+import { Card, IconButton, Text, useTheme } from "react-native-paper";
 import { monthToNumberMap } from '../util';
 
-
-// const yearlyPoints = [{
-//   yearlyPointsId: 1,
-//   contractId: 1,
-//   year: 2023,
-//   amountAvailable: 154,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 2,
-//   contractId: 1,
-//   year: 2024,
-//   amountAvailable: 200,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 3,
-//   contractId: 1,
-//   year: 2025,
-//   amountAvailable: 200,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 4,
-//   contractId: 2,
-//   year: 2023,
-//   amountAvailable: 0,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 5,
-//   contractId: 2,
-//   year: 2024,
-//   amountAvailable: 150,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 6,
-//   contractId: 2,
-//   year: 2025,
-//   amountAvailable: 150,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 7,
-//   contractId: 3,
-//   year: 2023,
-//   amountAvailable: 0,
-//   amountBanked: 50,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 8,
-//   contractId: 3,
-//   year: 2024,
-//   amountAvailable: 150,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }, {
-//   yearlyPointsId: 9,
-//   contractId: 3,
-//   year: 2025,
-//   amountAvailable: 100,
-//   amountBanked: 0,
-//   amountBorrowed: 0
-// }]
-
-export default function UseYearComponent({ contractsForUseYear, useYear }) {
+export default function UseYearComponent({ contractsForUseYear, useYear, handleDeleteContract }) {
 
   const theme = useTheme();
 
@@ -116,14 +50,23 @@ export default function UseYearComponent({ contractsForUseYear, useYear }) {
       fontSize: 16,
       color: theme.colors.primary,
     },
+    removeButtonRow: {
+      display: 'flex',
+      flexDirection: 'flex',
+      justifyContent: 'right',
+      alignContent: 'flex-end'
+    },
+    removeButton: {
+      margin: 5,
+      marginLeft: 'auto'
+    },
+    removeButtonLabel: {
+      color: 'red'
+    }
   });
 
-  useEffect(() => {
-
-  }, [contractsForUseYear])
-
   const buildAvailablePointsBlock = (contract) => {
-    const { contractId, useYear, points: fullPoints } = contract
+    const { contract_id, use_year, points: fullPoints } = contract
     const currentDate = new Date();
     const currentMonth = currentDate.getMonth()
     const currentYear = currentDate.getFullYear()
@@ -140,7 +83,7 @@ export default function UseYearComponent({ contractsForUseYear, useYear }) {
     while (yearToDisplay <= lastYearToDisplay) {
       // fetch yearly point for contract and year
 
-      const foundYearlyPoints = contract.allotments.filter(yearlyPoint => yearlyPoint.contractId === contractId && yearlyPoint.year === yearToDisplay)[0]
+      const foundYearlyPoints = contract.allotments.filter(yearlyPoint => yearlyPoint.contract_id === contract_id && yearlyPoint.year === yearToDisplay)[0]
       if (foundYearlyPoints != undefined) {
         availablePoints.push(foundYearlyPoints);
       }
@@ -148,9 +91,9 @@ export default function UseYearComponent({ contractsForUseYear, useYear }) {
     }
     return (
       availablePoints?.map((availablePoint, index) => (
-      <React.Fragment key={index}>
-        <View style={styles.yearRow}><Text style={styles.yearLabel}>{availablePoint.year}</Text><Text style={styles.availablePoints}>{`${availablePoint.points_available} pts`}</Text></View>
-      </React.Fragment>
+        <React.Fragment key={index}>
+          <View style={styles.yearRow}><Text style={styles.yearLabel}>{availablePoint.year}</Text><Text style={styles.availablePoints}>{`${availablePoint.points_available} pts`}</Text></View>
+        </React.Fragment>
       )))
   }
 
@@ -160,9 +103,17 @@ export default function UseYearComponent({ contractsForUseYear, useYear }) {
         <Text style={styles.useYearLabel}>{useYear}</Text>
         {contractsForUseYear?.map((contract, index) => (
           <React.Fragment key={index}>
-            <View style={styles.resortRow}><Text style={styles.homeResortLabel}>{contract.homeResort}</Text><Text style={styles.contractPoints}>{`${contract.points} pts`} </Text></View>
+            <View style={styles.resortRow}>
+              <Text style={styles.homeResortLabel}>{contract.homeResort}</Text>
+              <Text style={styles.contractPoints}>{`${contract.points} pts`} </Text>
+            </View>
             <View style={styles.availablePoints}>
               {buildAvailablePointsBlock(contract)}
+            </View>
+            <View style={styles.removeButtonRow}>
+              <TouchableOpacity style={styles.removeButton} onPress={() => handleDeleteContract(contract.contract_id, contract.use_year)}>
+                <Text style={styles.removeButtonLabel}>Remove</Text>
+              </TouchableOpacity>
             </View>
           </React.Fragment>
         ))}
